Allow floating point time zone

+++ src/PowderBot/Data/Models/UserModel.cs
@@ -30,7 +30,7 @@ namespace Data.Models
         public string Id { get; set; }
         public int NotifyAfter { get; set; }
         public int NotifyBefore { get; set; }
-        public int Gmt { get; set; }
+        public float Gmt { get; set; }
         public string LastCommand { get; set; }
     }
 }
